Multiple changes across many pages
Fixed a variety of issues across the pages and probably added several as well.
In registration.jsp the login credentials for the database are now correct so this *should* work as intended.

The login page should also work now, however I can't test either of these as MySQL won't work on my Mac currently with external programs that aren't the root user.

The db_init.sql file was updated with several changes including that I added a piece for every customer entry that they should have loyalty points assigned as the loyalty points would be assigned to the user itself and not the booking.

I updated the loyaltypoints.jsp page to check if the user is logged in first (since a logged in user would have 0 points) and then I filled out a (mostly semi-functional) db query but I can't figure out how to return the data and actually display it on screen for the user (the data being the person's first name and their loyalty points amount). Finally, I tried messing with index.jsp for as long as I could and I cannot for the life of me figure out how to make the "is the user logged in, if so, do they want to log out?" button that appears above the nav bar work right. I want it to have a "Sign in?" button that links to login.jsp if the user is not  signed in, and if the user is signed in I want it to welcome them by their first name and give them an option to logout, and so far I can't get either of these to work, I just cannot figure out the code for it.